+++ app/classes/web/base_handler.py
@@ -230,6 +230,19 @@ class BaseHandler(tornado.web.RequestHandler):
             api_key, token_data, user = self.controller.authentication.check_err(
                 self._auth_get_api_token()
             )
+            if not user["enabled"]:
+                # Log the user out and send error if disabled
+                self.clear_cookie("token")
+                return self.finish_json(
+                    403,
+                    {
+                        "status": "error",
+                        "error": "ACCESS_DENIED",
+                        "error_data": self.helper.translation.translate(
+                            "login", "accountDisabled", user["lang"]
+                        ),
+                    },
+                )
             if token_data.get("token_id") and token_data.get(
                 "session_id"
             ):  # Check to see if token defines session_id and token_id